orchestrator/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "arc-orchestrator"
version = "0.1.0"
description = "Goose-powered orchestrator for ARC Protocol — runs real agent sessions wired into the ARC MCP server and produces genuine signed provenance records."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [{ name = "ARC Protocol Contributors" }]
keywords = ["arc", "goose", "mcp", "agent", "orchestrator", "bitcoin"]
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "pydantic>=2.6.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "apscheduler>=3.10.0",
    "websockets>=12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "respx>=0.21.0",
]

[project.scripts]
arc-orchestrator = "arc_orchestrator.api:run"

[tool.hatch.build.targets.wheel]
packages = ["src/arc_orchestrator"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
